backend/requirements.txt
Flask
python-dotenv

# test
pytest

# production
gunicorn
